requirements.txt
numpy 
pandas
PyQt5
Pillow
geopy
difflib
pgeocode
matplotlib
